refactor(autonomi): use archive_public instead of _private
