+++ autonomi/src/client/files/mod.rs
@@ -1,8 +1,8 @@
 pub mod archive;
-pub mod archive_private;
+pub mod archive_public;
 #[cfg(feature = "fs")]
 #[cfg_attr(docsrs, doc(cfg(feature = "fs")))]
 pub mod fs;
 #[cfg(feature = "fs")]
 #[cfg_attr(docsrs, doc(cfg(feature = "fs")))]
-pub mod fs_private;
+pub mod fs_public;
