+++ autonomi/tests/external_signer.rs
@@ -5,8 +5,8 @@ use alloy::providers::Provider;
 use ant_evm::{QuoteHash, TxHash};
 use ant_logging::LogBuilder;
 use autonomi::client::external_signer::encrypt_data;
-use autonomi::client::files::archive::Metadata;
-use autonomi::client::files::archive_private::PrivateArchive;
+use autonomi::client::files::archive::PrivateArchive;
+use autonomi::client::files::archive_public::Metadata;
 use autonomi::client::payment::Receipt;
 use autonomi::client::vault::user_data::USER_DATA_VAULT_CONTENT_IDENTIFIER;
 use autonomi::client::vault::VaultSecretKey;
